+++ error-prone-contrib/src/main/java/tech/picnic/errorprone/refasterrules/ImmutableListRules.java
@@ -263,84 +263,4 @@ final class ImmutableListRules {
       return ImmutableList.of(e1, e2, e3, e4, e5);
     }
   }
-
-  /**
-   * Prefer {@link Stream#of(Object[])} over constructing a list and then directly calling {@link
-   * ImmutableList#stream()} on it.
-   */
-  static final class StreamImmutableListOf1<T> {
-    @BeforeTemplate
-    Stream<T> before(T e1) {
-      return ImmutableList.of(e1).stream();
-    }
-
-    @AfterTemplate
-    Stream<T> after(T e1) {
-      return Stream.of(e1);
-    }
-  }
-
-  /**
-   * Prefer {@link Stream#of(Object[])} over constructing a list and then directly calling {@link
-   * ImmutableList#stream()} on it.
-   */
-  static final class StreamImmutableListOf2<T> {
-    @BeforeTemplate
-    Stream<T> before(T e1, T e2) {
-      return ImmutableList.of(e1, e2).stream();
-    }
-
-    @AfterTemplate
-    Stream<T> after(T e1, T e2) {
-      return Stream.of(e1, e2);
-    }
-  }
-
-  /**
-   * Prefer {@link Stream#of(Object[])} over constructing a list and then directly calling {@link
-   * ImmutableList#stream()} on it.
-   */
-  static final class StreamImmutableListOf3<T> {
-    @BeforeTemplate
-    Stream<T> before(T e1, T e2, T e3) {
-      return ImmutableList.of(e1, e2, e3).stream();
-    }
-
-    @AfterTemplate
-    Stream<T> after(T e1, T e2, T e3) {
-      return Stream.of(e1, e2, e3);
-    }
-  }
-
-  /**
-   * Prefer {@link Stream#of(Object[])} over constructing a list and then directly calling {@link
-   * ImmutableList#stream()} on it.
-   */
-  static final class StreamImmutableListOf4<T> {
-    @BeforeTemplate
-    Stream<T> before(T e1, T e2, T e3, T e4) {
-      return ImmutableList.of(e1, e2, e3, e4).stream();
-    }
-
-    @AfterTemplate
-    Stream<T> after(T e1, T e2, T e3, T e4) {
-      return Stream.of(e1, e2, e3, e4);
-    }
-  }
-
-  /**
-   * Prefer {@link Stream#of(Object[])} over constructing a list and then directly calling {@link
-   * ImmutableList#stream()} on it.
-   */
-  static final class StreamImmutableListOf5<T> {
-    @BeforeTemplate
-    Stream<T> before(T e1, T e2, T e3, T e4, T e5) {
-      return ImmutableList.of(e1, e2, e3, e4, e5).stream();
-    }
-
-    @AfterTemplate
-    Stream<T> after(T e1, T e2, T e3, T e4, T e5) {
-      return Stream.of(e1, e2, e3, e4, e5);
-    }
-  }
 }
